Extracted rest.keystore from core.keystore bundle
