+++ kura/distrib/config/kura.build.properties
@@ -121,6 +121,7 @@ org.eclipse.kura.rest.security.provider.version=2.0.0-SNAPSHOT
 org.eclipse.kura.rest.identity.provider.version=2.0.0-SNAPSHOT
 org.eclipse.kura.rest.service.listing.provider.version=2.0.0-SNAPSHOT
 org.eclipse.kura.rest.system.provider.version=2.0.0-SNAPSHOT
+org.eclipse.kura.rest.keystore.provider.version=1.0.0-SNAPSHOT
 org.eclipse.kura.request.handler.jaxrs.version=2.0.0-SNAPSHOT
 org.eclipse.kura.rest.wire.provider.version=2.0.0-SNAPSHOT
 org.eclipse.kura.wire.ai.component.provider.version=2.0.0-SNAPSHOT
