+++ kura/org.eclipse.kura.core.keystore/META-INF/MANIFEST.MF
@@ -6,11 +6,6 @@ Bundle-Version: 2.0.0.qualifier
 Bundle-Vendor: Eclipse Kura
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 Import-Package: com.eclipsesource.json;version="0.9.5",
- com.google.gson;version="2.7.0",
- com.google.gson.annotations;version="2.7.0",
- javax.annotation.security;version="1.2.0",
- javax.ws.rs;version="2.0.1",
- javax.ws.rs.core;version="2.0.1",
  org.bouncycastle.asn1;version="1.78.1",
  org.bouncycastle.asn1.cms;version="1.78.1",
  org.bouncycastle.asn1.pkcs;version="1.78.1",
@@ -34,14 +29,8 @@ Import-Package: com.eclipsesource.json;version="0.9.5",
  org.bouncycastle.util.io.pem;version="1.78.1",
  org.eclipse.kura;version="[1.6,2.0)",
  org.eclipse.kura.certificate;version="[2.1,3.0)",
- org.eclipse.kura.cloudconnection.message;version="[1.0,2.0)",
- org.eclipse.kura.cloudconnection.request;version="[1.0,2.0)",
  org.eclipse.kura.configuration;version="[1.2,2.0)",
  org.eclipse.kura.crypto;version="[1.1,2.0)",
- org.eclipse.kura.marshalling;version="[1.0,2.0)",
- org.eclipse.kura.message;version="[1.4,2.0)",
- org.eclipse.kura.request.handler.jaxrs;version="[1.0,2.0)",
- org.eclipse.kura.rest.utils;version="[1.0,2.0)",
  org.eclipse.kura.security.keystore;version="[1.1,1.2)",
  org.eclipse.kura.system;version="[1.5,2.0)",
  org.eclipse.kura.util.configuration;version="[1.0,2.0)",
