--- kura/org.eclipse.kura.core.keystore/OSGI-INF/org.eclipse.kura.core.keystore.rest.provider.keystoreRestServiceV1.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-   Copyright (c) 2021, 2023 Eurotech and/or its affiliates and others
-
-   This program and the accompanying materials are made
-   available under the terms of the Eclipse Public License 2.0
-   which is available at https://www.eclipse.org/legal/epl-2.0/
- 
-	SPDX-License-Identifier: EPL-2.0
-	
-	Contributors:
-	 Eurotech
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" enabled="true" immediate="true" name="org.eclipse.kura.core.keystore.rest.provider.KeystoreRestServiceV1">
-   <implementation class="org.eclipse.kura.core.keystore.rest.provider.KeystoreRestServiceV1"/>
-
-	<property name="service.pid" value="org.eclipse.kura.core.keystore.rest.provider.KeystoreRestServiceV1"/>
- <reference bind="setUserAdmin" cardinality="1..1" interface="org.osgi.service.useradmin.UserAdmin" name="UserAdmin" policy="static"/>
- <service>
-    <provide interface="org.eclipse.kura.core.keystore.rest.provider.KeystoreRestService"/>
- </service>
-</scr:component>
