--- kura/org.eclipse.kura.core.keystore/src/main/java/org/eclipse/kura/core/keystore/request/handler/KeystoreServiceRequestHandlerV1.java
@@ -1,22 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Eurotech and/or its affiliates and others
- *
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *  Eurotech
- *  Red Hat Inc
- *******************************************************************************/
-package org.eclipse.kura.core.keystore.request.handler;
-
-public class KeystoreServiceRequestHandlerV1 extends KeystoreServiceRequestHandler {
-
-    public KeystoreServiceRequestHandlerV1() {
-        super("KEYS-V1");
-    }
-
-}
