+++ kura/org.eclipse.kura.rest.keystore.provider/OSGI-INF/org.eclipse.kura.internal.rest.keystore.request.handler.KeystoreServiceRequestHandlerV2.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+   Copyright (c) 2023 Eurotech and/or its affiliates and others
+
+   This program and the accompanying materials are made
+   available under the terms of the Eclipse Public License 2.0
+   which is available at https://www.eclipse.org/legal/epl-2.0/
+ 
+	SPDX-License-Identifier: EPL-2.0
+	
+	Contributors:
+	 Eurotech
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" enabled="true" immediate="true" name="org.eclipse.kura.internal.rest.keystore.request.handler.KeystoreRequestHandlerV2">
+   <implementation class="org.eclipse.kura.internal.rest.keystore.request.handler.KeystoreServiceRequestHandlerV2"/>
+
+	<property name="service.pid" value="org.eclipse.kura.internal.rest.keystore.request.handler.KeystoreServiceRequestHandlerV2"/>
+ <reference bind="setRequestHandlerRegistry" cardinality="0..n" interface="org.eclipse.kura.cloudconnection.request.RequestHandlerRegistry" name="RequestHandlerRegistry" policy="dynamic" unbind="unsetRequestHandlerRegistry"/>
+</scr:component>
