+++ kura/org.eclipse.kura.rest.keystore.provider/build.properties
@@ -1,0 +1,19 @@
+#
+#  Copyright (c) 2024 Eurotech and/or its affiliates and others
+#
+#  This program and the accompanying materials are made
+#  available under the terms of the Eclipse Public License 2.0
+#  which is available at https://www.eclipse.org/legal/epl-2.0/
+#
+#  SPDX-License-Identifier: EPL-2.0
+#
+#  Contributors:
+#   Eurotech
+#
+output.. = target/classes
+bin.includes = .,\
+               META-INF/,\
+               OSGI-INF/,\
+               about.html,\
+               about_files/
+source.. = src/main/java/
