+++ kura/org.eclipse.kura.rest.keystore.provider/src/main/java/org/eclipse/kura/internal/rest/keystore/provider/KeystoreRestServiceV1.java
@@ -1,0 +1,20 @@
+/*******************************************************************************
+ * Copyright (c) 2023, 2024 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *  Eurotech
+ *******************************************************************************/
+package org.eclipse.kura.internal.rest.keystore.provider;
+
+import javax.ws.rs.Path;
+
+@Path("/keystores/v1")
+public class KeystoreRestServiceV1 extends KeystoreRestService {
+
+}
