+++ kura/test/org.eclipse.kura.rest.keystore.provider.test/META-INF/MANIFEST.MF
@@ -1,0 +1,34 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: org.eclipse.kura.rest.keystore.provider.test
+Bundle-SymbolicName: org.eclipse.kura.rest.keystore.provider.test;singleton:=true
+Bundle-Version: 6.0.0.qualifier
+Bundle-Vendor: Eclipse Kura
+Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
+Bundle-ClassPath: .
+Bundle-ActivationPolicy: lazy
+Import-Package: com.eclipsesource.json;version="0.9.5",
+ javax.servlet;version="3.1.0",
+ javax.servlet.http;version="3.1.0",
+ org.apache.commons.io;version="2.11.0",
+ org.bouncycastle.asn1.x500;version="1.78.1",
+ org.eclipse.kura.configuration;version="[1.2,2.0)",
+ org.eclipse.kura.core.crypto,
+ org.eclipse.kura.core.keystore.util;version="[1.0,2.0)",
+ org.eclipse.kura.core.testutil;version="[1.0,2.0)",
+ org.eclipse.kura.core.testutil.http;version="[1.0,2.0)",
+ org.eclipse.kura.core.testutil.pki;version="[1.1.0,2.0.0)",
+ org.eclipse.kura.core.testutil.requesthandler;version="1.2.0",
+ org.eclipse.kura.core.testutil.service;version="1.0.0",
+ org.eclipse.kura.security.keystore;version="[1.0,2.0)",
+ org.junit;version="[4.12.0,5.0.0)",
+ org.junit.runners;version="[4.12.0,5.0.0)",
+ org.mockito;version="[4.0.0,5.0.0)",
+ org.mockito.invocation;version="[4.0.0,5.0.0)",
+ org.mockito.stubbing;version="[4.0.0,5.0.0)",
+ org.slf4j;version="1.6.4"
+Fragment-Host: org.eclipse.kura.rest.keystore.provider
+Require-Bundle: org.eclipse.kura.http.server.manager;bundle-version="1.4.0",
+ org.junit,
+ org.eclipse.kura.broker.artemis.core;bundle-version="1.5.0",
+ org.eclipse.kura.broker.artemis.simple.mqtt;bundle-version="1.4.0"
