+++ kura/test/org.eclipse.kura.rest.keystore.provider.test/build.properties
@@ -1,0 +1,20 @@
+#
+# Copyright (c) 2017, 2020 Eurotech and/or its affiliates and others
+# 
+# This program and the accompanying materials are made
+# available under the terms of the Eclipse Public License 2.0
+# which is available at https://www.eclipse.org/legal/epl-2.0/
+# 
+# SPDX-License-Identifier: EPL-2.0
+# 
+# Contributors:
+#  Eurotech
+#
+
+bin.includes = .,\
+               META-INF/,\
+               about.html
+source.. = src/main/java/
+additional.bundles = org.eclipse.kura.api,\
+                     slf4j.api,\
+                     org.apache.logging.log4j.api
