+++ kura/test/org.eclipse.kura.rest.keystore.provider.test/src/test/resources/cert
@@ -1,0 +1,16 @@
+-----BEGIN CERTIFICATE-----
+MIICezCCAjigAwIBAgIESvucGTALBgcqhkjOOAQDBQAwDzENMAsGA1UEAxMEa3Vy
+YTAeFw0xNzA4MjQxNDMxMjNaFw0zMTA1MDMxNDMxMjNaMA8xDTALBgNVBAMTBGt1
+cmEwggG3MIIBLAYHKoZIzjgEATCCAR8CgYEA/X9TgR11EilS30qcLuzk5/YRt1I8
+70QAwx4/gLZRJmlFXUAiUftZPY1Y+r/F9bow9subVWzXgTuAHTRv8mZgt2uZUKWk
+n5/oBHsQIsJPu6nX/rfGG/g7V+fGqKYVDwT7g/bTxR7DAjVUE1oWkTL2dfOuK2HX
+Ku/yIgMZndFIAccCFQCXYFCPFSMLzLKSuYKi64QL8Fgc9QKBgQD34aCF1ps93su8
+q1w2uFe5eZSvu/o66oL5V0wLPQeCZ1FZV4661FlP5nEHEIGAtEkWcSPoTCgWE7fP
+CTKMyKbhPBZ6i1R8jSjgo64eK7OmdZFuo38L+iE1YvH7YnoBJDvMpPG+qFGQiaiD
+3+Fa5Z8GkotmXoB7VSVkAUw7/s9JKgOBhAACgYAaqPBZ1sLviuLYcBbVYhtJRUB1
+FBrCxMut8sWDAFCEaZ/gLuOZP7Odw9ifuDLkfF9iDi20MNUbojgFLg1u71IOLUtW
+6+0vICqfZRai7i2kNHM82J1TLaS4QYQzTDGUxs0xAOSaidyBlGPvwrzTQixIK5Ph
+H0VBeEr9a2RIbaHgOKMhMB8wHQYDVR0OBBYEFJe86lxIv5UYsZ/uGcIdewCPUeaj
+MAsGByqGSM44BAMFAAMwADAtAhUAiw2QQPqB75io+p1cxQfMQW4Z/6oCFEA7AGYv
+kLQlC6I0o6bJqtKRlwar
+-----END CERTIFICATE-----
